backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
aiomysql
alembic
pydantic-settings
python-dotenv
openai
python-multipart
PyPDF2
structlog
